% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_kh.R
\name{get_kh}
\alias{get_kh}
\title{Wrapper function to import and format Krummh\enc{ö}{oe}rn data for children and families from SPSS files}
\usage{
get_kh(source_path = NULL, source_files = NULL, get_fam = FALSE)
}
\arguments{
\item{source_path}{File path to  the SPSS file.}

\item{source_files}{File names (ususally 'kindermc.sav' and 'familienmc.sav', as default)}

\item{Should}{family data be returned? Default is FALSE}
}
\description{
This function returns a data.frame containing genealogical information of individuala
}
\examples{
twinmat <- get_twinmat()
str(twinmat)
}
\keyword{import}
\keyword{kh}

