% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kh_evn.R
\docType{data}
\name{kh_evn}
\alias{kh_evn}
\title{Krummhörn Events (Long Format)}
\format{Data frame in long format}
\source{
see `citation("kh.data")`
}
\usage{
kh_evn
}
\description{
Data frame in long format providing variables in  dim(kh_mat)$vv .
}
\keyword{datasets}

