% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kh_geo.R
\docType{data}
\name{kh_geo}
\alias{kh_geo}
\title{Krummhörn Geo Positional Data}
\format{A list}
\source{
see `citation("kh.data")`
}
\usage{
kh_geo
}
\description{
List of three elements
[[1]] "df" Dataframe holding geocodes for valid locations
[[2]] "maps" A list of maps (as returned from ggmap::ggmap())
[[1]] "dist" Distance matrix derived from 'df'
}
\keyword{datasets}

