% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kh_ind.R
\docType{data}
\name{kh_ind}
\alias{kh_ind}
\title{Krummhörn Individuals}
\format{A dataframe}
\source{
see `citation("kh.data")`
}
\usage{
kh_ind
}
\description{
Dataframe for individual dates and genealogical relations.
}
\keyword{datasets}

