% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kh_kin.R
\docType{data}
\name{kh_kin}
\alias{kh_kin}
\title{Krummhörn Kin Tables}
\format{Nested list of dataframes}
\source{
see `citation("kh.data")`
}
\usage{
kh_kin
}
\description{
Nested lists of dataframes holding information on known pedigree members
being alive at each specific date
}
\keyword{datasets}

