% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kh_ped.R
\docType{data}
\name{kh_ped}
\alias{kh_ped}
\title{Krummhörn Pedigree List}
\format{A list of pedigrees}
\source{
see `citation("kh.data")`
}
\usage{
kh_ped
}
\description{
A list of pedigrees for 2572 mothers in families being classified as 'fully documented' (see  sample_kh ),
as returned from grap_ped(x) and named after the ID of each specific mother.
}
\keyword{datasets}

